data: sort data files after downloading them
The index construction process fairly deeply assumes that we visit
records in the IMDb data in lexicographic order of the `tt` identifiers.
Previously, the data was pre-sorted in this fashion. Unfortunately, the
addition of 10 character identifiers appears to break this property, or
more precisely, `tt1234567`, for example, would appear *after*
`tt12345670`, even though it is lexicographically smaller.

To fix this, we read all CSV data into memory when initially downloading
it, and then sort the data by the first two columns, since some data files
have identifiers in the second column as well. This is regrettable, but
is a simple fix.